server_model/requirements.txt
# Server Model Requirements
# PyTorch + EfficientNet + FastAPI

# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# ML utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Export formats
onnx>=1.14.0
onnxruntime>=1.15.0
# onnxruntime-gpu>=1.15.0  # Uncomment for GPU support

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
